"fixed" the gpio io issue + removed some code from registers
TURNS OUT THERE WASN'T AN ISSUE AT ALL.

PD1 is shared with SWIM, which is an active low communication
protocol. I had the programmer connected to the SWIM pin, which has a
separate header. When the SWIM cable is connected, the programmer
brings that pin high, causing the LED I was using to try and verify if
the pull-up was working to turn on. GPIO code is fully functional.

Doesn't quite explain the peripheral dump being inaccurate, but
honestly I'll take that at this point.